// web/src/Enums/RouteNames.ts
export enum RouteNames {
  RegistreraAbonnemangPage = "/registrera/abonnemang",
  RegistreraBredbandPage = "/registrera/bredband",
  RegistreraTvPage = "/registrera/tv",
  RegistreringskoderPage = "/registreringskoder",
  MobilDealPage = "/mobildeal",
  LathundPage = "/operatorspriser",
  VaruhusPage = "/varuhus",
  OrdrarPage = "/ordrar",
  OrderPage = "/order/:id",
  UnauthorizedPage = "/unauthorized",
}